Fix bad match in map_arity
It currently matches an improper list with the atom `head` as the head
and the atom `tail` as the tail. Erlang variables (bindings) start with
uppercase, so this looks like a typo.

Instead of a `case` with such a match we can use `if` and check the type
of `Arity`, which is excatly what this commit does.